// packages/daemon/src/generated/signalr/Source.Features.RuntimeBootstrap.Contracts.ts
export type RuntimeSpecV2 = {
    /** Transpiled from int */
    version: number;
    /** Transpiled from string? */
    install?: string;
    /** Transpiled from string? */
    installVerify?: string;
    /** Transpiled from System.Collections.Generic.List<Source.Features.RuntimeBootstrap.Contracts.ServiceSpec>? */
    services?: ServiceSpec[];
    /** Transpiled from string? */
    setup?: string;
}

/** Transpiled from Source.Features.RuntimeBootstrap.Contracts.ServiceSpec */
export type ServiceSpec = {
    /** Transpiled from string */
    name: string;
    /** Transpiled from string */
    command: string;
    /** Transpiled from string? */
    user?: string;
    /** Transpiled from bool */
    autorestart?: boolean;
    /** Transpiled from System.Collections.Generic.Dictionary<string, string>? */
    env?: Partial<Record<string, string>>;
    /** Transpiled from Source.Features.RuntimeBootstrap.Contracts.HealthcheckSpec? */
    healthcheck?: HealthcheckSpec;
    /** Transpiled from string? */
    install?: string;
    /** Transpiled from string? */
    installVerify?: string;
}

/** Transpiled from Source.Features.RuntimeBootstrap.Contracts.HealthcheckSpec */
export type HealthcheckSpec = {
    /** Transpiled from string */
    command: string;
    /** Transpiled from int */
    intervalSeconds?: number;
}